requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.3.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.34.0
azure-data-tables==12.7.0
azure-functions==1.23.0
azure-functions-durable==1.3.2
azure-identity==1.23.0
azure-storage-blob==12.25.1
beautifulsoup4==4.13.4
boto3==1.39.0
botocore==1.39.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
cryptography==45.0.4
distro==1.9.0
et-xmlfile==2.0.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fasta2a==0.3.5
fastavro==1.11.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
furl==2.1.4
google-auth==2.40.3
google-genai==1.23.0
googleapis-common-protos==1.70.0
griffe==1.7.3
groq==0.29.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
logfire==3.22.0
logfire-api==3.21.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.10.1
mdurl==0.1.2
mistralai==1.9.1
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.3
numpy==2.2.6
oauthlib==3.3.1
openai==1.93.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orderedmultidict==1.0.1
packaging==25.0
pandas==2.3.0
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.3.5
pydantic-ai-slim==0.3.5
pydantic-core==2.33.2
pydantic-evals==0.3.5
pydantic-graph==0.3.5
pydantic-settings==2.10.1
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.25.2
pypdfium2==4.30.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.1
structlog==25.4.0
tenacity==8.5.0
tokenizers==0.21.2
tomli==2.2.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
